Bug fix: use new input tokens to make requests
Previously we were using the old stored tokens to make the api calls
to update repository variables & secrets when using `set-arg` to input
a new token. This was fine as long as the keys being replaced were still
valid. In the case that the old key expried, it would fail to update
the corresponding var in the remote repo.

This commit separates the responsibility into two functions _ensuring_
that the new input tokens will be used.